requirements.txt
# math
numpy==1.26.4
scipy
scikit-learn
einops
seaborn
gym==0.23.1

jupyterlab 
jupyterlab_widgets
ipywidgets

# gymnax
gymnasium
matplotlib 
tqdm 
# brax
# dm-acme[jax]
orbax
ray[tune]
flax==0.8.4
tensorflow==2.15.1
tensorflow-datasets
rlds
dm_pix
dm-haiku
dm_reverb
tensorflow_addons
tensorflow_probability==0.23.0

opencv-python
moviepy

# Logging
wandb
ipdb
loguru

absl-py
dataclasses
ml_collections

# Env
xminigrid
# procgen-mirror # for python 3.11 and 3.12
procgen-mirror # for python 3.9

# Formatting
ruff

# Config
hydra-core
omegaconf

# Atari
tyro
gym[atari]
autorom[accept-rom-license]
stable_baselines3

jaxtyping
distrax

# d4rl
# 'Cython<3'